+++ lib/shrine/storage/file_system.rb
@@ -91,7 +91,7 @@ class Shrine
 
       # Deletes the specified directory on the filesystem.
       #
-      #    file_system.delete_prefixed("somekey/derivatives")
+      #    file_system.delete_prefixed("somekey/derivatives/")
       def delete_prefixed(delete_prefix)
         FileUtils.rm_rf directory.join(delete_prefix)
       end