+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* `core` – Add ability to call `Storage#delete_prefixed` with a prefix with trailing slash (@jrochkind)
+
 * `core` – Add `Shrine::Attachment[]` shorthand for `Shrine::Attachment.new` (@janko)
 
 ## 3.0.0.rc (2019-09-28)